chore(lib): move lib code in new folder structure
