+++ core/src/main/kotlin/io/github/bluegroundltd/outbox/item/OutboxPayload.kt
@@ -1,0 +1,3 @@
+package io.github.bluegroundltd.outbox.item
+
+interface OutboxPayload
