+++ core/src/main/kotlin/io/github/bluegroundltd/outbox/item/OutboxStatus.kt
@@ -1,0 +1,8 @@
+package io.github.bluegroundltd.outbox.item
+
+enum class OutboxStatus {
+  PENDING,
+  RUNNING,
+  COMPLETED,
+  FAILED
+}
