+++ core/src/main/kotlin/io/github/bluegroundltd/outbox/item/OutboxType.kt
@@ -1,0 +1,5 @@
+package io.github.bluegroundltd.outbox.item
+
+interface OutboxType {
+  fun getType(): String
+}
