+++ core/src/test/groovy/io/github/bluegroundltd/outbox/OutboxPayloadTest.groovy
@@ -1,0 +1,6 @@
+package io.github.bluegroundltd.outbox
+
+import io.github.bluegroundltd.outbox.item.OutboxPayload
+
+@SuppressWarnings("EmptyClass")
+class OutboxPayloadTest implements OutboxPayload {}
