+++ core/src/test/groovy/io/github/bluegroundltd/outbox/utils/UnitTestSpecification.groovy
@@ -1,0 +1,9 @@
+package io.github.bluegroundltd.outbox.utils
+
+import spock.lang.Specification
+
+class UnitTestSpecification extends Specification implements SpecHelper {
+  def setupSpec() {
+    TimeZone.setDefault(TimeZone.getTimeZone("UTC"))
+  }
+}
