--- lib/src/main/kotlin/transactional/outbox/Library.kt
@@ -1,10 +1,0 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package transactional.outbox
-
-class Library {
-    fun someLibraryMethod(): Boolean {
-        return true
-    }
-}
