--- modules/transactional-outbox/src/main/kotlin/com/blueground/outbox/OutboxHandler.kt
@@ -1,19 +1,0 @@
-package com.blueground.outbox
-
-import com.blueground.outbox.item.OutboxPayload
-import com.blueground.outbox.item.OutboxType
-import java.time.Instant
-
-interface OutboxHandler {
-  fun getSupportedType(): OutboxType
-
-  fun serialize(payload: OutboxPayload): String
-
-  fun getNextExecutionTime(currentRetries: Long): Instant
-
-  fun hasReachedMaxRetries(retries: Long): Boolean
-
-  fun handle(payload: String)
-
-  fun handleFailure(payload: String)
-}
