--- modules/transactional-outbox/src/main/kotlin/com/blueground/outbox/TransactionalOutbox.kt
@@ -1,12 +1,0 @@
-package com.blueground.outbox
-
-import com.blueground.outbox.item.OutboxPayload
-import com.blueground.outbox.item.OutboxType
-
-// TODO document both interface and each method (expected functionality)
-interface TransactionalOutbox {
-
-  fun add(type: OutboxType, payload: OutboxPayload)
-
-  fun monitor()
-}
