--- modules/transactional-outbox/src/main/kotlin/com/blueground/outbox/item/OutboxPayload.kt
@@ -1,3 +1,0 @@
-package com.blueground.outbox.item
-
-interface OutboxPayload
