--- modules/transactional-outbox/src/main/kotlin/com/blueground/outbox/item/OutboxStatus.kt
@@ -1,8 +1,0 @@
-package com.blueground.outbox.item
-
-enum class OutboxStatus {
-  PENDING,
-  RUNNING,
-  COMPLETED,
-  FAILED
-}
