--- modules/transactional-outbox/src/main/kotlin/com/blueground/outbox/item/OutboxType.kt
@@ -1,5 +1,0 @@
-package com.blueground.outbox.item
-
-interface OutboxType {
-  fun getType(): String
-}
