--- modules/transactional-outbox/src/test/groovy/com/blueground/outbox/OutboxPayloadTest.groovy
@@ -1,5 +1,0 @@
-package com.blueground.outbox
-
-import com.blueground.outbox.item.OutboxPayload
-
-class OutboxPayloadTest implements OutboxPayload {}
