--- modules/transactional-outbox/src/test/groovy/com/blueground/outbox/utils/UnitTestSpecification.groovy
@@ -1,9 +1,0 @@
-package com.blueground.outbox.utils
-
-import spock.lang.Specification
-
-class UnitTestSpecification extends Specification implements SpecHelper {
-  def setupSpec() {
-    TimeZone.setDefault(TimeZone.getTimeZone("UTC"))
-  }
-}
